requirements.txt
cachetools==4.1.0
decorator==4.3.0
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.9.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.3
opencv-python==4.2.0.34
packaging==20.3
Pillow==7.1.1
requests==2.23.0
requests-file==1.4.3
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
streamlit==0.58.0
tensorflow==2.0.0
smtplib